[FIX] TasteNovelGetResponse DTO의 interestCount 타입 변경
from Integer to Long

+++ src/main/java/org/websoso/WSSServer/dto/userNovel/TasteNovelGetResponse.java
@@ -8,7 +8,7 @@ public record TasteNovelGetResponse(
         String title,
         String author,
         String novelImage,
-        Integer interestCount
+        Long interestCount
 //        Float novelRating,
 //        Integer novelRatingCount
 ) {
